Telemetry for metadata type/component count for beta deploy/retrieve @W-7759892@ (#2299)
